fix: memperbaiki dari repository
memulai dari basic dan next akan menambahkan beberapa action serta listing ke issue untuk materi basic
