+++ README.md
@@ -1,6 +1,10 @@
 # Kotlin
 Repositori ini berisi kumpulan dari berbagai macam contoh struktur data, algoritma dan komputasi matematika yang diimplementasikan dengan menggunakan bahasa pemrograman Kotlin.
 
+## Tentang Kotlin
+Kotlin merupakan bahasa pemograman yang diperkenalkan oleh JetBrains pada tahun 2011. jetbrains adalah perancang resmi dari _Intergrated Development Environment_(IDE) Java yang yang 
+sangat canggih, yang dikenal dengan intellij IDEA. Kotlin adalah bahasa yang grates dan akan tetap gratis. bahasa ini diekembangkan dibawah lisensi [Apache 2.0](https://www.apache.org/licenses/LICENSE-2.0) dan kode sumbernya dapat diakses melalui [github](https://github.com/JetBrains/kotlin). Kotlin adalah bahasa pemrograman umum yang menggunakan tipe data statis yang kuat dan berjalan di atas Java Virtual Machine (JVM). Pada tahun 2017, Google mengumumkan Kotlin sebagai bahasa resmi untuk pengembangan aplikasi Android. Kotlin juga merupakan bahasa pemrograman sumber terbuka yang menggabungkan fitur-fitur pemrograman berorientasi objek dan fitur-fitur fungsional dalam platform yang unik.
+
 ## Peraturan dan Pedoman Kontribusi
 Dipersilahkan bila ingin melakukan open pull request atau kontribusi pada repositori **kotlin**. Sebelum itu, bacalah terlebih dahulu [peraturan dan pedomannya](CONTRIBUTING.md) yang dapat ditemukan pada ``CONTRIBUTING.md``.
