--- basic/helloWorld/README.md
@@ -1,11 +1,0 @@
-# Hello World
-
-Titik utama dari aplikasi kotlin adalah pada fungsi ``main()``. Sebuah fungsi dapat didefinisikan sebagai blok kode yang dirancang
-dan melakukan tugas tertentu.
-
-```kotlin
-fun main() {
-    print("bellshade")
-}
-```
-kamu juga bisa mengganti dari fungsi ``println()`` ke fungsi ``print()``. Fungsi ``print()`` dalam Kotlin digunakan untuk menampilkan argumennya pada output standar, sementara fungsi ``println()`` melakukan tugas yang sama namun juga menambahkan pemisah baris pada output.
