--- basic/helloWorld/helloWorld.kt
@@ -1,12 +1,0 @@
-/*
-  Titik utama dari aplikasi kotlin adalah pada fungsi ``main()``. 
-  Sebuah fungsi dapat didefinisikan sebagai blok kode yang dirancang
-  dan melakukan tugas tertentu.
-*/
-fun main() {
-  // menampilkan kata ``bellshade``
-
-  // kamu bisa menggunakan fungsi dari print() atau println()
-  //print("bellshade")
-  println("bellshade")
-}
