--- basic/main/00-hello-kotlin/HelloKotlin.kt
@@ -1,8 +1,0 @@
-// Ini adalah kode dasar menggunakan "Hello World" menggunakan kotlin!
-
-fun main() {
-
-  // Di kotlin, kita tidak wajib menuliskan ';' (titik koma)  di ujung kode
-  // Hal ini karena kotlin telah mengenali ujung dari kode itu sendiri
-  print("Hello, Kotlin!")
-}
