--- basic/main/01-variable-dan-tipe-data/Boolean.kt
@@ -1,12 +1,0 @@
-fun main(){
-  /**
-   * Tipe data boolean dideklarasikan dengan keyword `Boolean`. 
-   * Valuenya hanya boleh `true` atau `false` 
-   */
-  val bool1 : Boolean = true
-  println(bool1)
-
-  // atau bisa hanya dengan
-  val bool2 = true
-  println(bool2)
-}
