--- basic/main/01-variable-dan-tipe-data/Char.kt
@@ -1,19 +1,0 @@
-fun main(){
-
-  /**
-   * Tipe data char dideklarasikan dengan keyword char. 
-   * Char harus dideklarasikan dengan single quotes (' '). 
-   * Kita juga bisa menggunakan unicode pada char. 
-   * 
-   */
-  val char1 : Char = 'A'
-  println(char1)
-
-  // atau bisa hanya dengan
-  val char2 = 'A'
-  println(char2)
-
-  // contoh penggunaan unicode saat deklarasi char
-  val char3 = '\u0041'
-  println(char3)
-}
