+++ src/main/kotlin/Basic/HelloWorld/HelloWorld.kt
@@ -1,0 +1,14 @@
+package Basic.HelloWorld
+
+/*
+  Titik utama dari aplikasi kotlin adalah pada fungsi ``main()``.
+  Sebuah fungsi dapat didefinisikan sebagai blok kode yang dirancang
+  dan melakukan tugas tertentu.
+*/
+fun main() {
+    // menampilkan kata ``bellshade``
+
+    // kamu bisa menggunakan fungsi dari print() atau println()
+    //print("bellshade")
+    println("bellshade")
+}
